UrlSecuriser/role/user/supprimer/comment/evenement

+++ src/Controller/CommentaireController.php
@@ -3,38 +3,38 @@
 namespace App\Controller;
 
 
+use App\Entity\User;
 use App\Entity\Evenement;
+
 use App\Entity\Commentaire;
-use App\Repository\CommentaireRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Security;
 
 class CommentaireController extends AbstractController
 {
 
 
-/**
- * @Route("/commentaire/{id}", name="supprimer_commentaire")
- */
-public function supprimerCommentaire(ManagerRegistry $doctrine, Commentaire $commentaire,SessionInterface $session)
-{
- 
-  
+    /**
+     * @Route("/commentaire/{id}", name="supprimer_commentaire")
+     */
+    public function supprimerCommentaire(ManagerRegistry $doctrine, Commentaire $commentaire = null, SessionInterface $session)
+    {
+if($commentaire){
+        if ($commentaire->getUtilisateur() == $this->getUser()) {
 
-    // Récupérer l'ID de l'événement pour rediriger l'utilisateur vers la page de détails de l'événement
-    $evenementId = $commentaire->getEvenement()->getId();
+            // Récupérer l'ID de l'événement pour rediriger l'utilisateur vers la page de détails de l'événement
+            $evenementId = $commentaire->getEvenement()->getId();
+            $entityManager = $doctrine->getManager();
+            $entityManager->remove($commentaire);
+            $entityManager->flush();
 
-    $entityManager = $doctrine->getManager();
-    $entityManager->remove($commentaire);
-    $entityManager->flush();
-
-    $session->getFlashBag()->add('success', 'Le commentaire a été supprimé avec succès.');
-
-    return $this->redirectToRoute('details_evenement', ['id' => $evenementId]);
+            $session->getFlashBag()->add('success', 'Le commentaire a été supprimé avec succès.');
+            return $this->redirectToRoute('details_evenement', ['id' => $evenementId]);
+        }
+    }
+        return $this->redirectToRoute('app_categorie');
+    }
 }
-
-
-}
